Patron id or barcode handling for ItemRequested
This is the second of two issues trying to fully align a fork with this main repo. ItemRequested messages send the barcode in VisibleUserIdentifier. To place the hold the patron id then needs to be retrieved using the barcode. Sorry, I don't have direct access for testing any longer so this is a proposed, but untested solution.